Updated the loan payment email facility

+++ Loan-Management-System/src/main/java/com/aurionpro/lms/dto/LoanSchemeRequestDTO.java
@@ -1,5 +1,7 @@
 package com.aurionpro.lms.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -10,7 +12,7 @@ public class LoanSchemeRequestDTO {
 	private String schemeName;
 
 	@NotNull(message = "Interest rate is required")
-	private Double interestRate;
+	private BigDecimal interestRate;
 
 	@NotNull(message = "Tenure in months is required")
 	private Integer tenureMonths;
