+++ Loan-Management-System/src/main/java/com/aurionpro/lms/dto/LoanSchemeResponseDTO.java
@@ -1,12 +1,14 @@
 package com.aurionpro.lms.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class LoanSchemeResponseDTO {
 	private int id;
 	private String schemeName;
-	private Double interestRate;
+	private BigDecimal interestRate;
 	private Integer tenureMonths;
 	private int adminId; // The admin who created this scheme
 }
